Move package sampling to processor/tailsampling
